Change the Service Endpoint for KAS to utilize HTTP2
This work is for https://gitlab.com/gitlab-org/gitlab/-/issues/365945

We're trying to resolve an issue where currently inside of our GKE
infrastructure, websockets and proxying commands to/from a Kubernetes
cluster is not usable.  This is an experiment to see if this resolves
our issues.

This resolution includes swapping over our Service Object to speak to
our services using HTTP2 vs the default HTTP1.1 version.  As this is a
managed Google infrastructure configuration, this is configured via an
annotation at the service level.  As port 8150 is the responding port
for these connections, we target this by the port name.

This is preprod so there is little harm if this is a failed experiment,
we'll simply revert if needed.

Reference: https://cloud.google.com/kubernetes-engine/docs/how-to/ingress-http2