Remove the last references to `only` and `except` from all of CI
Part of https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/2410

This change is to remove all final references to `only:` and `except:`
from all our CI setup.

The overall goal is for the pipeline structure to be the exact same
as it was before, this is simply a synxtax change to remove the
final refernces to these deprecated CI stanzas.