Change all use of `when: always` to `when: on_success`
Fixes https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/2579

According to the documentation `when: always` will always run the
job in question regardless of failures in the previous stage. We
should be using `when: on_success` to mimic the normal GitLab behavior
of only running when all jobs in the previous stage succeed.