Move to using helm log for monitoring rollout of resources
Corrective action for https://gitlab.com/gitlab-com/gl-infra/production/-/issues/7808

Previously we weren't able to get a realtime log from helm using helmfile, which
meant we were not able to leverage helms built in logging to show what the helm
processes is doing and/or monitoring for the rollout of the resources it was updating.

The recent addition of https://github.com/helmfile/helmfile/pull/286 to helmfile
now allows us to access the helm log in realtime.

This change drops the `watch-gitlab-deployments.sh` script we currently use
to monitor Kubernetes resource rollout, which is a quite a hack (sometimes doesn't
work, and doesn't work at all for canary) with the actual helm log which also
shows us what resources helm thinks have changed, and what it's actually waiting on
to mark the helm release as complete (or roll it back).

This should be far more accurate and less error prone overall.